requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
streamlit==1.29.0
tqdm==4.66.1
openai==1.55.3
python-dotenv==1.0.0
plotly==5.18.0
httpx==0.24.1
